add email notification
