+++ novu/bridge/app/novu/workflows/uploader/index.ts
@@ -1,1 +1,1 @@
-export * from "./commit-acknowledged";
+export * from "./commitAcknowledged";
